zerko-interview-agent/requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-community
langchain-google-genai
pydantic
python-dotenv
requests
pymupdf
prisma 
asyncio
tenacity

